start migrating helper tests
